fix: use fp32 dtype for Florence-2 encoder — fp16 ONNX incompatible with WASM backend
The fp16 quantized ONNX files use SimplifiedLayerNormFusion nodes that
ONNX Runtime Web's WASM backend doesn't support. Switched encoder,
vision encoder, and embed_tokens to fp32. Decoder stays q4 for size.
Also added console.error to the scan catch block for debugging.

+++ lib/receipt-ocr.ts
@@ -40,9 +40,9 @@ async function getModel(onProgress?: (p: OcrProgress) => void) {
     const [model, processor, tokenizer] = await Promise.all([
       Florence2ForConditionalGeneration.from_pretrained(MODEL_ID, {
         dtype: {
-          embed_tokens: 'fp16',
-          vision_encoder: 'fp16',
-          encoder_model: 'fp16',
+          embed_tokens: 'fp32',
+          vision_encoder: 'fp32',
+          encoder_model: 'fp32',
           decoder_model_merged: 'q4',
         },
         progress_callback: (p: any) => {
